Move OntologyTermType to Core API.
